requirements.txt
customtkinter>=5.2.0
psutil>=5.9.0
Pillow>=9.0.0
